requirements.txt
Django==2.2.3
django-cors-headers==3.0.2
django-filter==2.1.0
django-rest-framework-mongoengine==3.3.1
djangorestframework==3.9.4
Markdown==3.1.1
mongoengine==0.18.2
numpy==1.16.4
pandas==0.24.0
pymongo==3.8.0
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
